Insert PerNodeTaskBuilder logic into PerNodeWorkloadGenerator.
